Crate paiagram_ui

Definitions for the user interface.

Modules

actions 🔒

A set of actions the user can perform Each action has a reverse action that would be triggered when the user hits the revert shortcut

command_palette 🔒

export_typst_diagram 🔒

save 🔒

tabs 🔒

widgets 🔒

Macros

for_all_tab_types 🔒

for_all_tabs 🔒

Structs

AdditionalTabViewer 🔒

AdditionalUiState 🔒

CoordinateSelection 🔒

ExtendingRouteSelection 🔒

ExtendingTripSelection 🔒

FrameTimeHistory 🔒

GlobalTimer 🔒

IntervalSelection 🔒

MainTabViewer 🔒

MainUiState 🔒

OpenOrFocus 🔒

StationSelection 🔒

TripSelection 🔒

UiModal 🔒

UiPlugin

Enums

AdditionalTab 🔒

MainTab 🔒

Modals 🔒

ModifySelectedItems 🔒

SelectedItem 🔒

SelectedItems 🔒

Functions

apply_custom_fonts

build_font_definitions 🔒

load_sarasa_local 🔒

open_or_focus_tab 🔒

show_name_button 🔒

show_ui

sync_ui 🔒

update_selected_items 🔒

update_timer 🔒