paiagram_ui

Module actions

A set of actions the user can perform Each action has a reverse action that would be triggered when the user hits the revert shortcut

Modules

change_entry_mode 🔒

Macros

for_all_actions 🔒

Structs

ActionHistory 🔒

ActionsPlugin 🔒

Enums

RevertableActions 🔒

Traits

RevertableAction 🔒