paiagram_ui::tabs

Trait Tab

pub trait Tab: MapEntities {
    const NAME: &'static str;

    // Required method
    fn main_display(&mut self, world: &mut World, ui: &mut Ui);

    // Provided methods
    fn edit_display(&mut self, _world: &mut World, ui: &mut Ui) { ... }
    fn display_display(&mut self, _world: &mut World, ui: &mut Ui) { ... }
    fn export_display(&mut self, _world: &mut World, ui: &mut Ui) { ... }
    fn title(&self) -> WidgetText { ... }
    fn on_tab_button(&self, _world: &mut World, _response: &Response) { ... }
    fn id(&self) -> Id { ... }
    fn scroll_bars(&self) -> [bool; 2] { ... }
    fn frame(&self) -> Frame { ... }
    fn icon(&self) -> Cow<'static, str> { ... }
}

Required Associated Constants

const NAME: &'static str

The internal name of the tab used for identification. This must be a static string. The actual displayed name could be different based on e.g. the localization or other contents.

Required Methods

fn main_display(&mut self, world: &mut World, ui: &mut Ui)

The main display of the tab.

Provided Methods

fn edit_display(&mut self, _world: &mut World, ui: &mut Ui)

The edit display

fn display_display(&mut self, _world: &mut World, ui: &mut Ui)

The details display

fn export_display(&mut self, _world: &mut World, ui: &mut Ui)

The export display

fn title(&self) -> WidgetText

The title of the tab

fn on_tab_button(&self, _world: &mut World, _response: &Response)

What to do when clicking on the tab button

fn id(&self) -> Id

The id of the tab

fn scroll_bars(&self) -> [bool; 2]

Whether if the tab allows scrolling

fn frame(&self) -> Frame

The frame of the tab

fn icon(&self) -> Cow<'static, str>

The icon of the tab

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.